chat/requirements.txt
openai==1.55.3
pymongo==4.10.1
python-dotenv==1.0.1
